+++ src/loss/bce.rs
@@ -29,6 +29,6 @@ impl BinaryCrossEntropyLoss {
                     .sum::<Value>()
             })
             .sum::<Value>()
-            / (ypred.len() * ypred[0].len()) as f32
+            / (ypred.len() * ypred[0].len()) as f64
     }
 }
