+++ src/loss/ce.rs
@@ -24,6 +24,6 @@ impl CrossEntropyLoss {
                     .sum::<Value>()
             })
             .sum::<Value>()
-            / (ypred.len() * ypred[0].len()) as f32
+            / (ypred.len() * ypred[0].len()) as f64
     }
 }
